requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pillow>=8.0.0
streamlit>=1.0.0
numpy>=1.19.0
matplotlib>=3.3.0
opencv-python>=4.5.0
